finish osworld baseline
