--- agential/agents/mm_agents/llm_server/CogAgent/README.md
@@ -1,7 +1,0 @@
-## Deploy CogAgent as server
-
-```
-python CogAgent.py
-```
-
-The CogAgent LLM will be deployed on http://127.0.0.1:8000
